Make it clearer that the main thread is not blocked on MediaNavigator construction
